requirements.txt
box2d-py==2.3.8
pygame==2.6.1
setuptools==75.6.0
wheel==0.45.1